DOMA-1138 move UnitNameInfo component to property domain, rebase with master
